+++ packages/terra-core-docs/src/terra-dev-site/test/list/InstructionTextList.test.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import List, { Item } from 'terra-list/lib/index';
+
+const InstructionTextList = () => (
+  <List instructionText="Here are some instructions for using this list.">
+    <Item key="0">
+      <p>Item 1</p>
+    </Item>
+    <Item key="1">
+      <p>Item 2</p>
+    </Item>
+    <Item key="2">
+      <p>Item 3</p>
+    </Item>
+  </List>
+);
+
+export default InstructionTextList;