+++ packages/terra-list/tests/jest/List.test.jsx
@@ -64,6 +64,22 @@ it('should render with thin padded items', () => {
   expect(shallowComponent).toMatchSnapshot();
 });
 
+it('should render with thin padded items', () => {
+  const item1 = <Item key="123" />;
+  const item2 = <Item key="124" />;
+  const item3 = <Item key="125" />;
+  const item4 = <Item key="126" />;
+  const item5 = <Item key="127" />;
+  const items = [item1, item2, item3, item4, item5];
+  const shallowComponent = shallow(
+    <div>
+      <p id="list-help"> Navigate this list using the arrow keys. </p>
+      <List instructions="list-help">{items}</List>
+    </div>,
+  );
+  expect(shallowComponent).toMatchSnapshot();
+});
+
 it('correctly applies the theme context className', () => {
   const wrapper = mount(
     <ThemeContextProvider theme={{ className: 'orion-fusion-theme' }}>